Pattern design + Junit + HighScore + Bugfixes

+++ src/main/java/uet/project/arkanoid/objects/brickVariants/IndestructibleBrick.java
@@ -6,7 +6,7 @@ import uet.project.arkanoid.objects.Brick;
 
 
 public class IndestructibleBrick extends Brick {
-    public IndestructibleBrick(int x, int y, double width, double height, double rotation, GameSetup stage) {
+    public IndestructibleBrick(double x, double y, double width, double height, double rotation, GameSetup stage) {
         super(x, y, width, height, rotation, 10, BrickType.INDESTRUCTIBLE, stage);
         this.brickImage = Basis.BRICK_INDESTRUCTIBLE_TEXTURE;
     }
